añadidos controladores para usuarios y login

+++ controllers/blogs.js
@@ -1,9 +1,12 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog.js')
-const logger = require('../utils/logger.js')
+const User = require('../models/user.js')
+const userExtractor = require('../middlewares/userExtractor.js')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({})
+    .populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
@@ -12,18 +15,35 @@ blogsRouter.get('/:id', async (request, response) => {
   response.json(blog)
 })
 
-blogsRouter.post('/', async (request, response) => {
-  const blog = request.body
+blogsRouter.post('/', userExtractor, async (request, response, next) => {
+  try {
+    const blog = request.body
 
-  const newBlog = new Blog({
-    title: blog.title,
-    author: blog.author,
-    url: blog.url,
-    likes: 0
-  })
+    const { userID } = request
 
-  const blogSaved = await newBlog.save()
-  response.status(201).json(blogSaved)
+    const user = await User.findById(userID)
+
+    if (!user) {
+      return response.status(401).json({
+        error: 'You must be logged in to create a blog'
+      })
+    }
+
+    const newBlog = new Blog({
+      title: blog.title,
+      author: blog.author,
+      url: blog.url,
+      likes: 0,
+      user: user._id
+    })
+
+    const blogSaved = await newBlog.save()
+    user.blogs = user.blogs.concat(blogSaved._id)
+    await user.save()
+    response.status(201).json(blogSaved)
+  } catch (e) {
+    next(e)
+  }
 })
 
 blogsRouter.put('/:id', async (request, response, next) => {
@@ -38,9 +58,26 @@ blogsRouter.put('/:id', async (request, response, next) => {
   response.status(201).json(updatedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', userExtractor, async (request, response, next) => {
+  try {
+    const { userID } = request
+    const user = await User.findById(userID)
+    const blog = await Blog.findById(request.params.id)
+
+    console.log({ user })
+    console.log({ blog })
+
+    if (user._id.toString() !== blog.user.toString()) {
+      return response.status(401).json({
+        error: 'You are not authorized to delete this blog'
+      })
+    }
+
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } catch (e) {
+    next(e)
+  }
 })
 
 module.exports = blogsRouter
